setup both sql and nosql database with prisma
