+++ .gitignore
@@ -138,3 +138,4 @@ dist
 
 
 funturf-schema.md
+backend/src/generated/
