+++ backend/prisma/postgresql/schema.prisma
@@ -1,0 +1,9 @@
+generator client {
+    provider = "prisma-client-js"
+    output   = "../../src/generated/prisma/pg"
+}
+
+datasource db {
+    provider = "postgresql"
+    url      = env("POSTGRESQL_DATABASE_URL")
+}
