+++ backend/src/prisma.js
@@ -1,4 +1,4 @@
-import {PrismaClient} from "./generated/prisma/client.js"
+import {PrismaClient} from "./generated/prisma/mongo/client.js"
 
 const prisma = new PrismaClient()
